Change how time is updated when going back steps
Time spent on a step is now always attributed to that step, even if you
then go back a step, rather than the time being attributed to the step
you go back to.

Going back after the task is finished still attributes the time after
finishing to the last step.